Move slerp into a separate file for license purposes
